[monit] Fix monit reload when running on localhost
Monit's reload command returns *before* it has actually finished
reloading, so we need to sleep. This wasn't an issue running remotely,
because the ssh latency was enough.

We can't use the built-in service module's sleep parameter either,
because that only works on *restart*, and not *reload* (which is what we
want).

Special thanks and apologies to @nanpanman for finding and reporting all
these Windows-setup and localhost-based bugs.