[tor] Add CPU nice support
Via a new `tor_nice` option. My server, droidsans.benjam.info, was
sustaining ~100mbit (great!), but eating a lot of CPU power in the
process. As a preventative measure (to avoid pissing off the VPS
provider), I wanted to control that a little better, and nice is easy to
add, since the existing init.d script supports it.

Unlike bitcoind, cpulimit is still probably a bit less useful for this,
because CPU and memory are the only real resources being consumed. With
bitcoind, we can use cpulimit as a poor-man's ionice on OpenVZ. I'll add
cpulimit if this continues to be an issue.